Fix of NPE in button by moving method implementations. Added tint ability for floating as workaround. Added warning logs.
